zmiany w dokumentacji
* dodano requirements.txt
* uzupełnienie pliku README o kilka sekcji
* Zagraj.bat uruchamia grę bez potrzeby korzystania z IDE
* przeniesiono menu.py do katalogu głównego
