+++ README.md
@@ -1,5 +1,14 @@
 # Mastermind_AiR_4b Dokumentacja
 
+## Spis treści
+* [Funkcjonalości](#funkcjonalnoci)
+* [Wybór platformy](#wybr-platformy)
+* [Na czym polega Mastermind?](#na-czym-polega-mastermind)
+* [Technologia](#technologia)
+* [Przykład działania](#przykad-dziaania)
+* [Uruchamianie](#uruchamianie)
+* [Autorzy](#autorzy)
+
 ## Funkcjonalności
 - Implementacja zasad gry.
 - Gra z komputerem.
@@ -20,3 +29,29 @@ Przy tworzeniu projektu korzystano z następujących narzędzi:
  - Python 3.8
  - pygame 1.9.6
  - JSON
+ 
+ ## Przykład działania
+<details>
+  <summary>Rozwiń</summary>
+  
+ ![Tak to działa](/gfx/how_it_works.gif)
+ 
+</details>
+
+## Uruchamianie
+Aby uruchomić grę należy najpierw upewnić się, czy posiada się wszystkie wymagane pakiety.
+W tym celu będąc w katalogu głównym gry wpisujemy w cmd:
+'''
+pip install -r requirements.txt
+'''
+A następnie aby już zagrać:
+'''
+python Menu/menu.py
+'''
+Bądź też można skorzystać z pliku 'Zagraj.bat', który włączy grę za nas.
+
+## Autorzy
+- a
+- b
+- c
+- d
