+++ Zagraj.bat
@@ -1,0 +1,1 @@
+python menu.py
