+++ requirements.txt
@@ -1,0 +1,2 @@
+pygame==1.9.6
+regex==2020.7.14
